SOC-2898 | refactor: combine  email confirmation controllers
