+++ extensions/wikia/Email/Controller/EmailConfirmationController.class.php
@@ -5,9 +5,7 @@ namespace Email\Controller;
 use Email\EmailController;
 use Email\Check;
 
-class EmailConfirmationController extends EmailController {
-
-	const TYPE = "ConfirmationMail";
+abstract class AbstractEmailConfirmationController extends EmailController {
 
 	const LAYOUT_CSS = EmailController::NO_ADDITIONAL_STYLES;
 
@@ -37,6 +35,45 @@ class EmailConfirmationController extends EmailController {
 		$this->assertUserHasEmail();
 	}
 
+	/**
+	 * @template otherEmailConfirmation
+	 */
+	public function body() {
+		$this->response->setData( [
+			'salutation' => $this->getSalutation(),
+			'summary' => $this->getSummary(),
+			'buttonLink' => $this->confirmUrl,
+			'buttonText' => $this->getButtonText(),
+			'contentFooterMessages' => $this->getEmailSpecificFooterMessages(),
+			'signature' => $this->getMessage( 'emailext-emailconfirmation-community-team' )->text()
+		] );
+	}
+
+	abstract protected function getSummary();
+
+	protected function getButtonText() {
+		return $this->getMessage( 'emailext-emailconfirmation-button-text' )->text();
+	}
+
+	abstract protected function getEmailSpecificFooterMessages();
+
+	protected static function getEmailSpecificFormFields() {
+		return [
+			'inputs' => [
+				[
+					'type' => 'hidden',
+					'name' => 'confirmUrl',
+					'value' => "#",
+				],
+			]
+		];
+	}
+}
+
+class EmailConfirmationController extends AbstractEmailConfirmationController {
+
+	const TYPE = "ConfirmationMail";
+
 	protected function getSubject() {
 		return $this->getMessage( 'emailext-emailconfirmation-subject' )->text();
 	}
@@ -52,17 +89,13 @@ class EmailConfirmationController extends EmailController {
 			'buttonText' => $this->getButtonText(),
 			'contentFooterMessageTop' => $this->getContentFooterMessageTop(),
 			'contentFooterList' => $this->createContentFooterList(),
-			'contentFooterMessagesBottom' => $this->getContentFooterMessagesBottom(),
+			'contentFooterMessagesBottom' => $this->getEmailSpecificFooterMessages(),
 			'signature' => $this->getMessage( 'emailext-emailconfirmation-community-team' )->text()
 		] );
 	}
 
-	private function getSummary() {
+	protected function getSummary() {
 		return $this->getMessage( 'emailext-emailconfirmation-summary' )->text();
-	}
-
-	private function getButtonText() {
-		return $this->getMessage( 'emailext-emailconfirmation-button-text' )->text();
 	}
 
 	private function getContentFooterMessageTop() {
@@ -78,21 +111,98 @@ class EmailConfirmationController extends EmailController {
 		];
 	}
 
-	private function getContentFooterMessagesBottom() {
+	protected function getEmailSpecificFooterMessages() {
 		return [
 			$this->getMessage( 'emailext-emailconfirmation-footer-2' )->text()
 		];
 	}
+}
+
+class EmailConfirmationReminderController extends AbstractEmailConfirmationController {
+
+	protected function getSubject() {
+		return $this->getMessage( 'emailext-emailconfirmation-reminder-subject', $this->getTargetUserName() )->parse();
+	}
+
+	protected function getSummary() {
+		return $this->getMessage( 'emailext-emailconfirmation-reminder-summary' )->text();
+	}
+
+	protected function getEmailSpecificFooterMessages() {
+		return [
+			$this->getMessage( 'emailext-emailconfirmation-reminder-footer-1',
+				$this->getTargetUserName() )->parse()
+		];
+	}
+}
+
+class ConfirmationChangedEmailController extends AbstractEmailConfirmationController {
+
+	private $newEmail;
+
+	public function initEmail() {
+		parent::initEmail();
+
+		$this->newEmail = $this->request->getVal( 'newEmail' );
+		$this->assertValidChangedParams();
+	}
+
+	protected function assertValidChangedParams() {
+		if ( empty( $this->newEmail ) ) {
+			throw new Check( "A value must be passed for parameter 'newEmail'" );
+		}
+	}
+
+	protected function getTargetUserEmail() {
+		return $this->newEmail;
+	}
+
+	protected function getSubject() {
+		return $this->getMessage( 'emailext-emailconfirmation-changed-subject' )->text();
+	}
+
+	protected  function getSummary() {
+		return $this->getMessage( 'emailext-emailconfirmation-changed-summary' )->text();
+	}
+
+	protected function getEmailSpecificFooterMessages() {
+		return [
+			$this->getMessage( 'emailext-emailconfirmation-changed-footer-1' )->text(),
+			$this->getMessage( 'emailext-emailconfirmation-changed-footer-2' )->text(),
+		];
+	}
 
 	protected static function getEmailSpecificFormFields() {
+		$parentForm = parent::getEmailSpecificFormFields();
+
+		$parentForm['inputs'][] = [
+			'type' => 'text',
+			'name' => 'newEmail',
+			'label' => "New Email",
+			'tooltip' => "The user's new email",
+		];
+
+		return $parentForm;
+	}
+}
+
+class ReactivateAccountController extends AbstractEmailConfirmationController {
+
+	protected function getSubject() {
+		return $this->getMessage( 'emailext-reactivate-account-subject' )->text();
+	}
+
+	protected function getSummary() {
+		return $this->getMessage( 'emailext-reactivate-account-summary' )->text();
+	}
+
+	protected function getButtonText() {
+		return $this->getMessage( 'emailext-reactivate-account-button-text' )->text();
+	}
+
+	protected function getEmailSpecificFooterMessages() {
 		return [
-			'inputs' => [
-				[
-					'type' => 'hidden',
-					'name' => 'confirmUrl',
-					'value' => "#",
-				],
-			]
+			$this->getMessage( 'emailext-reactivate-account-welcome-back' )->text(),
 		];
 	}
 }
